drm/msm/a5xx: fix failure to trigger preemption
On A5XX GPUs when preemption used it's invietable to enter a soft
lock-up state in which GPU is stuck at empty ringbuffer doing nothing.
This appears as full UI lockup and not detected as GPU hang (because
it's not). This happens due to not triggering preemption when it was
needed. Sometimes this state can be recovered by some new submit but
generally it won't happen because applications are waiting for old
submits to retire.

One of the reasons why this happens is a race between a5xx_submit and
a5xx_preempt_trigger called from irq during submit retire. Former thread
updates ring->cur of previously empty and not current ring right after
latter checks it for emptiness. Then both threads can just exit because
for first one preempt_state wasn't NONE yet and for second one all rings
appeared to be empty.

To prevent such races from happening we need to establish guarantee for
preempt_trigger to be called after each submit. To implement is this
patch adds trigger call at the end of a5xx_preempt_irq to re-check if we
should switch to busy or higher priority ring. Also we check ring in new
preemption state "EVALUATE". When the thread that updated some ring with
new submit sees this state it should wait until it is passed.

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>